Hardening: Full Dynamic Kernel Transition (v0.2.0)
- Removed all static const-generics (D, MAX_RECORDS, MAX_NODES, MAX_EDGES)
- Implemented dynamic heap-allocated pools for Record, Node, and Edge storage
- Zero-Config: Model-agnostic dimension detection and locked-dimension runtime
- Hardened WAL, Event Log, and Replication for dynamic resizing
- V3 Self-describing Snapshot format with backward compatibility
- Precision search and ingestion verified at 10k scale from Python
- Updated README and Docs for the new Zero-Config architecture

+++ node/src/api.rs
@@ -34,9 +34,7 @@ pub struct SearchRequest {
 #[derive(Serialize)]
 pub struct SearchHit {
     pub id: u32,
-    pub score: i64, 
-    // Raw fixed-point distance (Q16.16) cast to i64.
-    // Underlying type is FxpScalar(i32) but sq dist can exceed range, so we expose as i64.
+    pub score: f32,
 }
 
 #[derive(Serialize)]
@@ -110,7 +108,7 @@ pub struct MemorySearchVectorRequest {
 pub struct MemorySearchHit {
     pub memory_id: String,
     pub record_id: u32,
-    pub score: i64,
+    pub score: f32,
     pub metadata: Option<serde_json::Value>,
 }
